requirements.txt
Flask==2.3.3
MetaTrader5>=5.0.37
pandas==2.1.1
numpy==1.26.0
pytz==2025.7
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==2.3.7
dnspython==2.4.2
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
Flask-Mail==0.10.0
